replace SCSceneCode and Logic Controllers

+++ Align2D/Services/StorageManager.swift
@@ -32,26 +32,49 @@ final class StorageManager {
         backgroundViewContext.automaticallyMergesChangesFromParent = true
     }
 
-    //MARK: - Core Data Methods
-    func fetch(completion: @escaping (Result<[VectorEntity], Error>) -> Void) {
-        let fetchRequest = VectorEntity.fetchRequest()
-        let sortDescriptor = NSSortDescriptor(key: "id", ascending: false)
-        fetchRequest.sortDescriptors = [sortDescriptor]
+        //MARK: - Core Data Methods
+        //    func fetch(completion: @escaping (Result<[VectorEntity], Error>) -> Void) {
+        //        let fetchRequest = VectorEntity.fetchRequest()
+        //        let sortDescriptor = NSSortDescriptor(key: "id", ascending: false)
+        //        fetchRequest.sortDescriptors = [sortDescriptor]
+        //
+        //        do {
+        //            let vectors = try self.backgroundViewContext.fetch(fetchRequest)
+        //            DispatchQueue.main.async {
+        //                completion(.success(vectors))
+        //            }
+        //        } catch {
+        //            DispatchQueue.main.async {
+        //                completion(.failure(error))
+        //            }
+        //        }
+        //    }
 
-        do {
-            let vectors = try self.backgroundViewContext.fetch(fetchRequest)
-            DispatchQueue.main.async {
-                completion(.success(vectors))
-            }
-        } catch {
-            DispatchQueue.main.async {
-                completion(.failure(error))
+    func fetch(completion: @escaping (Result<[VectorEntity], Error>) -> Void) {
+        backgroundViewContext.perform { [weak self] in
+            guard let self else { return }
+
+            let fetchRequest = VectorEntity.fetchRequest()
+            let sortDescriptor = NSSortDescriptor(key: "id", ascending: false)
+            fetchRequest.sortDescriptors = [sortDescriptor]
+
+            do {
+                let vectors = try self.backgroundViewContext.fetch(fetchRequest)
+                DispatchQueue.main.async {
+                    completion(.success(vectors))
+                }
+            } catch {
+                DispatchQueue.main.async {
+                    completion(.failure(error))
+                }
             }
         }
     }
 
     func saveVector(vector: VectorModel, completion: @escaping (Result<Void, Error>) -> Void) {
-        backgroundViewContext.perform {
+        backgroundViewContext.perform { [weak self] in
+            guard let self else { return }
+
             let entity = VectorEntity(context: self.backgroundViewContext)
             entity.id = vector.id
             entity.startX = Double(vector.startX)
@@ -62,16 +85,21 @@ final class StorageManager {
 
             do {
                 try self.backgroundViewContext.save()
-                completion(.success(()))
+                DispatchQueue.main.async {
+                    completion(.success(()))
+                }
             } catch {
                 self.backgroundViewContext.rollback()
-                completion(.failure(error))
+                DispatchQueue.main.async {
+                    completion(.failure(error))
+                }
             }
         }
     }
 
     func delete(vector: VectorEntity, completion: @escaping (Result<Void, Error>) -> Void) {
-        backgroundViewContext.perform {
+        backgroundViewContext.perform { [weak self] in
+            guard let self else { return }
             self.backgroundViewContext.delete(vector)
             do {
                 try self.backgroundViewContext.save()
@@ -86,11 +114,45 @@ final class StorageManager {
             }
         }
     }
+
+    func updateVector(id: UUID, startX: Double, startY: Double, endX: Double, endY: Double, completion: @escaping (Result<Void, Error>) -> Void) {
+        backgroundViewContext.perform { [weak self] in
+            guard let self = self else { return }
+
+            let request: NSFetchRequest<VectorEntity> = VectorEntity.fetchRequest()
+            request.predicate = NSPredicate(format: "id == %@", id as CVarArg)
+
+            do {
+                let results = try self.backgroundViewContext.fetch(request)
+
+                if let vector = results.first {
+                        // Обновляем координаты
+                    vector.startX = startX
+                    vector.startY = startY
+                    vector.endX = endX
+                    vector.endY = endY
+
+                    try self.backgroundViewContext.save()
+                    DispatchQueue.main.async {
+                        completion(.success(()))
+                    }
+                } else {
+                    DispatchQueue.main.async {
+                        completion(.failure(NSError(domain: "UpdateError", code: 404, userInfo: [NSLocalizedDescriptionKey: "Vector not found"])))
+                    }
+                }
+            } catch {
+                DispatchQueue.main.async {
+                    completion(.failure(error))
+                }
+            }
+        }
+    }
 }
 
     // MARK: - Core Data Saving support
 extension StorageManager {
-    
+
     func saveContext () {
         if viewContext.hasChanges {
             do {
